relbld: Add repertoire builder utility
This commit adds a repertoire builder utility, repbld, which
interactively prompts the user to specify an opening
repertoire. During construction, repbld queries the lichess openings
database in order to find the most common replies. Additionally repbld
makes use of the eval facility to show the user the evaluation of
various replies during construction.