repmk: add optional --mingames argument
This commit adds an optional --mingames argument allows the user to
specify the minimum number of games in an opening book from a given
position. When building a repertoire, repmk will cease deepening a
varation once this threshold is hit. Previously this value was
hardcoded to 200, which is now the default.